Add more entity configuration

+++ src/Falcata.BillPlanner.Domain/Models/BillPlanner/DebtPeriod/DebtPeriod.cs
@@ -5,7 +5,7 @@ namespace Falcata.BillPlanner.Domain.Models.BillPlanner.DebtPeriod;
 
 public class DebtPeriod: BaseEntity<long>
 {
-    public long AccountId { get; set; }
+    public int AccountId { get; set; }
     public DateTimeOffset CutOffDate { get; set; }
     public DateTimeOffset PayDeadlineDate { get; set; }
     public DateTimeOffset InitCutOffDate { get; set; }
